# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "OpenAPI service for managing Aliyun FC and Cloudflare R2.",
        "title": "OpenAPI Service",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/api/v1/services/aliases": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update service alias to point to a specific version",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "aliases"
                ],
                "summary": "Update alias",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer {token}",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Environment (e.g. pre, prod)",
                        "name": "Env",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Country code (e.g. US, CN)",
                        "name": "Country-Code",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Update alias request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.UpdateAliasRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "400": {
                        "description": "Invalid request headers or body",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get list of aliases for a specific service",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "aliases"
                ],
                "summary": "List aliases",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer {token}",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Environment (e.g. pre, prod)",
                        "name": "Env",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Country code (e.g. US, CN)",
                        "name": "Country-Code",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "List alias request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.ListAliasRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "400": {
                        "description": "Invalid request headers or body",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    }
                }
            }
        },
...
        "/api/v1/services/publish": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Publish a service version",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "services"
                ],
                "summary": "Publish service",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer {token}",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Environment (e.g. pre, prod)",
                        "name": "Env",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Country code (e.g. US, CN)",
                        "name": "Country-Code",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Publish service request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.PublicServiceRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "400": {
                        "description": "Invalid request headers or body",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handler.ListAliasRequest": {
            "type": "object",
            "required": [
                "ServiceName"
            ],
            "properties": {
                "ServiceName": {
                    "type": "string"
                }
            }
        },
...
        "handler.PublicServiceRequest": {
            "type": "object",
            "required": [
                "description",
                "serviceName"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "serviceName": {
                    "type": "string"
                }
            }
        },
        "handler.UpdateAliasRequest": {
            "type": "object",
            "required": [
                "aliasName",
                "serviceName",
                "versionId"
            ],
            "properties": {
                "aliasName": {
                    "type": "string"
                },
                "serviceName": {
                    "type": "string"
                },
                "versionId": {
                    "type": "string"
                }
            }
        },
...
        "model.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {},
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string",
                    "example": "Success"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}